requirements.txt
flask
flask-cors
pymongo
python-dotenv
beautifulsoup4
playwright
requests
